+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # CHANGELOG
 
+## Version 0.0.5
+Build: 191201
+
+### Updated
+
+- Add Text component
+- Add Textarea component
+- Update logo project and app :tada:
+- Update layout Example
+
+
+
 ## Version 0.0.4
 Build: 191130